Add tests for checkWinner and checkTie
